Add operator-sdk boilerplate
This commit introduces boilerplate as generated by operator-sdk.  It is
roughly the result of the following commands:

  operator-sdk init --domain=skydive.network --repo=github.com/openshift/skydive-operator
  operator-sdk create api --group config --version v1alpha1 --kind Skydive --resource=true --controller=true

and then I started adding a tiny bit of code in the skydive resource
controller, but it just gets the Skydive resource and does nothing with
it.

The first target is to create a simple working operator to
then iterate on.  The next commits will create a default Skydive
deployment when the Skydive resource exists.